[Cleanup] Remove fwdelc declarations of functions moved to .h/.cpp
This means lots of #includes were missing in files.
