+++ src/ESPEasy.ino
@@ -90,8 +90,6 @@
 // Plugin helper needs the defined controller sets, thus include after 'define_plugin_sets.h'
 #include "_CPlugin_Helper.h"
 
-
-
 #include "ESPEasyWiFi_credentials.h"
 #include "ESPEasyWifi.h"
 #include "ESPEasyWifi_ProcessEvent.h"
@@ -126,14 +124,14 @@
 #include "src/Globals/Statistics.h"
 
 #include "src/Helpers/DeepSleep.h"
+#include "src/Helpers/ESPEasyRTC.h"
+#include "src/Helpers/ESPEasy_FactoryDefault.h"
 #include "src/Helpers/ESPEasy_Storage.h"
 #include "src/Helpers/ESPEasy_checks.h"
-#include "src/Helpers/ESPEasyRTC.h"
 #include "src/Helpers/Hardware.h"
 #include "src/Helpers/Misc.h"
 #include "src/Helpers/PeriodicalActions.h"
 #include "src/Helpers/Scheduler.h"
-
 
 #if FEATURE_ADC_VCC
 ADC_MODE(ADC_VCC);
