+++ src/ESPEasyWifi_ProcessEvent.cpp
@@ -11,10 +11,13 @@
 #include "src/Globals/MQTT.h"
 #include "src/Globals/NetworkState.h"
 #include "src/Globals/RTC.h"
+#include "src/Helpers/ESPEasyRTC.h"
 #include "src/Helpers/ESPEasy_Storage.h"
 #include "src/Helpers/ESPEasy_time_calc.h"
+#include "src/Helpers/Misc.h"
 #include "src/Helpers/Scheduler.h"
 #include "src/Helpers/StringConverter.h"
+
 
 bool unprocessedWifiEvents() {
   if (processedConnect && processedDisconnect && processedGotIP && processedDHCPTimeout)
