+++ src/_C016.ino
@@ -25,6 +25,8 @@ The controller can deliver the data to:
 <TODO>
 */
 
+#include "src/DataStructs/ESPEasyControllerCache.h"
+
 #define CPLUGIN_016
 #define CPLUGIN_ID_016         16
 #define CPLUGIN_NAME_016       "Cache Controller [Experimental]"
