+++ src/_CPlugin_Helper.cpp
@@ -17,6 +17,7 @@
 
 #include "src/Helpers/CompiletimeDefines.h"
 #include "src/Helpers/ESPEasy_time_calc.h"
+#include "src/Helpers/Misc.h"
 #include "src/Helpers/StringConverter.h"
 
 #include <WiFiClient.h>
