+++ src/_Plugin_Helper.cpp
@@ -8,6 +8,7 @@
 #include "src/Globals/Plugins.h"
 #include "src/Globals/Settings.h"
 #include "src/Globals/SecuritySettings.h"
+#include "src/Helpers/Misc.h"
 
 
 PluginTaskData_base *Plugin_task_data[TASKS_MAX] = { nullptr, };
