+++ src/src/Commands/Common.cpp
@@ -2,10 +2,14 @@
 
 #include <ctype.h>
 #include <IPAddress.h>
+
+#include "../../ESPEasyWifi.h"
 #include "../../ESPEasy_common.h"
 #include "../../ESPEasy_fdwdecl.h"
+
 #include "../DataStructs/ESPEasy_EventStruct.h"
 #include "../DataStructs/EventValueSource.h"
+
 #include "../Helpers/Numerical.h"
 #include "../Helpers/StringConverter.h"
