+++ src/src/Commands/Diagnostic.cpp
@@ -7,27 +7,31 @@
  #include "../DataStructs/ESPEasy_EventStruct.h"
  */
 
+
+#include "../../ESPEasy_Log.h"
+#include "../../ESPEasy_fdwdecl.h"
+
+#include "../Commands/Common.h"
+
+#include "../DataStructs/PortStatusStruct.h"
 #include "../DataStructs/SettingsType.h"
+#include "../DataStructs/SettingsType.h"
+
+#include "../Globals/Device.h"
+#include "../Globals/ExtraTaskSettings.h"
+#include "../Globals/GlobalMapPortStatus.h"
+#include "../Globals/SecuritySettings.h"
+#include "../Globals/Settings.h"
+#include "../Globals/Statistics.h"
+
+#include "../Helpers/Convert.h"
+#include "../Helpers/ESPEasy_Storage.h"
+#include "../Helpers/ESPEasy_time_calc.h"
+#include "../Helpers/Misc.h"
+#include "../Helpers/StringConverter.h"
 
 #include <map>
 #include <stdint.h>
-
-#include "../Commands/Common.h"
-#include "../Globals/Settings.h"
-#include "../Globals/SecuritySettings.h"
-#include "../Globals/ExtraTaskSettings.h"
-#include "../Globals/Device.h"
-#include "../DataStructs/SettingsType.h"
-#include "../DataStructs/PortStatusStruct.h"
-#include "../Globals/GlobalMapPortStatus.h"
-#include "../../ESPEasy_Log.h"
-#include "../Globals/Statistics.h"
-#include "../Helpers/ESPEasy_Storage.h"
-#include "../Helpers/ESPEasy_time_calc.h"
-#include "../Helpers/StringConverter.h"
-#include "../Helpers/Convert.h"
-
-#include "../../ESPEasy_fdwdecl.h"
 
 
 #ifndef BUILD_MINIMAL_OTA
