+++ src/src/Commands/InternalCommands.cpp
@@ -34,6 +34,7 @@
 #include "../Commands/wd.h"
 #include "../Commands/WiFi.h"
 
+#include "../Helpers/Misc.h"
 #include "../Helpers/StringConverter.h"
