+++ src/src/Commands/Timer.cpp
@@ -1,13 +1,17 @@
 #include "../Commands/Timer.h"
 
 
-#include "../Commands/Common.h"
-#include "../../ESPEasy_fdwdecl.h"
-#include "../../ESPEasy_common.h"
 
 #include "../../ESPEasy-Globals.h"
+#include "../../ESPEasy_common.h"
+#include "../../ESPEasy_fdwdecl.h"
+
+#include "../Commands/Common.h"
+
 #include "../Globals/ESPEasy_Scheduler.h"
+
 #include "../Helpers/ESPEasy_time_calc.h"
+#include "../Helpers/Misc.h"
 #include "../Helpers/Scheduler.h"
 
 String command_setRulesTimer(int msecFromNow, int timerIndex, int recurringCount) {
