+++ src/src/Commands/UDP.cpp
@@ -6,6 +6,7 @@
 #include "../Commands/Common.h"
 #include "../Globals/NetworkState.h"
 #include "../Globals/Settings.h"
+#include "../Helpers/Misc.h"
 #include "../Helpers/StringConverter.h"
 
 String Command_UDP_Test(struct EventStruct *event, const char *Line)
