+++ src/src/DataStructs/ESPEasyControllerCache.h
@@ -1,5 +1,5 @@
-#ifndef HELPERS_ESPEASYCONTROLLERCACHE_H
-#define HELPERS_ESPEASYCONTROLLERCACHE_H
+#ifndef DATASTRUCTS_ESPEASYCONTROLLERCACHE_H
+#define DATASTRUCTS_ESPEASYCONTROLLERCACHE_H
 
 #include <Arduino.h>
 
@@ -48,4 +48,4 @@ private:
 };
 
 
-#endif // ifndef HELPERS_ESPEASYCONTROLLERCACHE_H
+#endif // ifndef DATASTRUCTS_ESPEASYCONTROLLERCACHE_H
