+++ src/src/Globals/Plugins.cpp
@@ -2,14 +2,18 @@
 
 #include "../../ESPEasy_Log.h"
 #include "../../_Plugin_Helper.h"
+
 #include "../DataStructs/ESPEasy_EventStruct.h"
 #include "../DataStructs/ESPEasy_plugin_functions.h"
 #include "../DataStructs/TimingStats.h"
+
 #include "../Globals/Device.h"
 #include "../Globals/ESPEasy_Scheduler.h"
 #include "../Globals/ExtraTaskSettings.h"
 #include "../Globals/GlobalMapPortStatus.h"
 #include "../Globals/Settings.h"
+
+#include "../Helpers/ESPEasyRTC.h"
 #include "../Helpers/ESPEasy_Storage.h"
 #include "../Helpers/Hardware.h"
 #include "../Helpers/Misc.h"
