+++ src/src/Helpers/ESPEasy_FactoryDefault.h
@@ -1,0 +1,11 @@
+#ifndef HELPERS_ESPEASY_FACTORYDEFAULT_H
+#define HELPERS_ESPEASY_FACTORYDEFAULT_H
+
+
+/********************************************************************************************\
+   Reset all settings to factory defaults
+ \*********************************************************************************************/
+void ResetFactory();
+
+
+#endif
