+++ src/src/Helpers/Misc.h
@@ -321,12 +321,6 @@ String getTaskDeviceName(taskIndex_t TaskIndex);
 
 
 /********************************************************************************************\
-   Reset all settings to factory defaults
- \*********************************************************************************************/
-void ResetFactory();
-
-
-/********************************************************************************************\
    If RX and TX tied together, perform emergency reset to get the system out of boot loops
  \*********************************************************************************************/
