+++ src/src/Helpers/PeriodicalActions.cpp
@@ -16,7 +16,9 @@
 #include "../Globals/SecuritySettings.h"
 #include "../Globals/Services.h"
 #include "../Globals/Statistics.h"
+#include "../Helpers/ESPEasyRTC.h"
 #include "../Helpers/Hardware.h"
+#include "../Helpers/Misc.h"
 
 
 /*********************************************************************************************\
