+++ src/src/PluginStructs/P015_data_struct.cpp
@@ -1,5 +1,6 @@
 #include "P015_data_struct.h"
 
+#include "../Helpers/Misc.h"
 
 #ifdef USES_P015
