+++ src/src/PluginStructs/P023_data_struct.cpp
@@ -1,5 +1,7 @@
 #include "P023_data_struct.h"
 
+#include "../Helpers/Misc.h"
+
 #ifdef USES_P023
 const char Plugin_023_myFont_Size[] PROGMEM = {
   0x05, // SPACE
