+++ src/src/PluginStructs/P044_data_struct.cpp
@@ -1,7 +1,8 @@
 #include "P044_data_struct.h"
 
-#include "../Helpers/ESPEasy_Storage.h"
 #include "../Globals/EventQueue.h"
+#include "../Helpers/ESPEasy_Storage.h"
+#include "../Helpers/Misc.h"
 
 #define P044_RX_WAIT              PCONFIG(0)
